scraper, July Update
Added a script to parse a month's worth of data.  bugHunt is just a test tool, modify as needed.  Fixed the fact that it was dropped postponed games with a dumb workaround; if the game is postponed, the necessary files don't exist.  So, if those files aren't found, we simply return None.  I also had to fix a bug where occasionally the start_speed of pitch didn't exist for no apparent reason; now it should check that it exists.  Also added timestamps for starting and finishing a day's worth of data.