getPlayers function
Getting pitcher and batter data was absurdly slow, so I added a new function, getPlayers() which uses the players.xml file in the main game directory.  It doesn't have some information the individual player files in the pitchers and batters folders did (no height and weight, for example), but only having to load one file should drastically speed up the process.  Also added an error handling protocol that was currently in the getPitchers() function to the main gameParse() function.  If the boxscore.xml file can't be opened, it'll print 'Game Postponed' and return None.

The gameParse() function hasn't been updated to use the new getPlayers() funciton; will get it in there soon.